feat(designtokens): add narrow fonts to design tokens
